Image Loading and File Management
Added a utility class for manipulating and creating folders and
directories. The project requirements dictate that the application must
run on the school computers, which run Windows 7. Thus, only Windows
suppor was considered.

Expanded upon the SFML implementation and the utiliy FileSystem class to
be able to autonomously load textures into a map with the path as the
key.