Add files via upload
Changed up the files so that it could run on windows!
Some notable changes ---

Actions now handled in the PlayerAction.cpp (Card checks are still handled by GUI/Buttons.cpp ) (Strategy)
Modified actions in the PlayerRoleCard.cpp (Decorator)
Set of actions are set to players when starting a new game. (corresponding the roleCardVal) Found in the board.cpp
Pllayer.cpp now has new attributes and methods. Including vector CARDSSELECTED to figure out the cards selected, to be used in the PlayerAction.cpp methods.
And another attribute CITYSELECTED, needed for CHARTERFLIGHT, as player has to spend the matching city card and get the option to fly anywhere. (NOT SURE HOW TO GET THE INPUT FROM GUI ISSUE #115)